Moved fadeAllStations to fade when return from map
