+++ frontend/src/components/Navbar/Navbar.js
@@ -1,12 +1,18 @@
 import './Navbar.css'
 import { Link } from "react-router-dom";
-
+import fadeAllStations from '../../fadeAllStations';
 
 const Navbar = ({stop, setTapInVisible}) => {
   const navigateAway = () => {
     stop();
     setTapInVisible(true);
   }
+
+  const fadeStationsWhenReturningToMap = () => {
+    console.log('fade stations when returning to map')
+    fadeAllStations();
+  }
+
   return (
     <nav>
       <div className="nav-container">
@@ -16,7 +22,7 @@ const Navbar = ({stop, setTapInVisible}) => {
           <img src="./Underground.png" alt="" className="luso-logo second-logo" />
         </div>
         <div className="nav-links">
-          {/data/.test(window.location.href) && <div className="map-link link-btn"><Link to="/sounds-of-the-underground">Map</Link></div>}
+          {/data/.test(window.location.href) && <div onClick={fadeStationsWhenReturningToMap} className="map-link link-btn"><Link to="/sounds-of-the-underground">Map</Link></div>}
           {/sounds/.test(window.location.href) && <div className="data-link link-btn"><Link to="/data">Data</Link></div>}
           <div className="exit-link link-btn"><Link to="/" onClick={navigateAway}>Exit</Link></div>
         </div>
