+++ frontend/src/components/SideBarLeft/SideBarLeft.css
@@ -15,7 +15,7 @@
   cursor: pointer;
 }
 :root {
-  --service-1: #ebf086;
+  --service-1: #eefaaf;
   --service-2: #b2e63a;
 }
